chapter14/requirements_jupyter.txt
pandas
numpy
matplotlib
mysql-connector 
mysql-connector-python 
 